(#3) Add validation provider for pushing packages
These changes implement a new validation handler that runs globally
on a Chocolatey CLI execution.
This validation handler makes it possible for us to validate any
implemented rule when a package is pushed to CCR, and if one or more
rules fails it will prevent any further execution of the push process to
continue.