(#6) Add empty description requirement rule
This adds a new rule for validating the requirement about a description
must be present, and it can not be empty. This is a duplicate of the
original CLI Rule CHCR0001 with more information and a link to the
actual documentation.

In the future we should look at a way to replace this existing rule, or
update the original rule to match this one so it can be removed from this
extension.