(#10) Add rule for requiring dependency on deprecated pkg
We want all deprecated packages to contain at least one dependency for
the package that will be superseded by deprecating this package.

This commit implements the same rule that is implemented in package
validator to make that happen by first checking if the title contains
the word `deprecated` and then checks if there are any dependencies
available.