(#19) Add action to build project
This adds a new GitHub workflow file to be able to build the project
when a PR or a community member pushes up changes.

This is to give a quicker indication on failures than having to look at
our private builds (or a community member needing to reach out to a
Chocolatey Team Member to check).

The artifacts from the build (the extension/packages) themself are not
included with this build as they won't be able to be ran on any systems.
Only the unit tests can be ran as part of this workflow.