(#26) Use milestone for license URL replacement
This updates the relpacement used when creating a license URL
to make use of the Build Version Milestone, instead of the tag name.

This is done as in certain scenarios the tag name is an empty string and will
cause an invalid URL to be generated. This causes problems when doing a
release of the product, and thus changing to the milestone is more
appropriate.