(#35) Implement note rule CPMR0062
This implements the note rule CPMR0062 - Chocolatey Dependency, which
checks if there is a dependency on Chocolatey CLI defined in the nuspec
file. This is checked as normally there is no need to define Chocolatey
CLI as a dependency, unless a specific feature is being used. As such,
it is considered to be an anti-pattern to define one.