(#34) Implement note rule CPMR0061 - Id Contains "."
This implements the note rule CPMR0061 that verifies whether a package
uses a dot in its identifier, which is not recommended to do unless
absolutely necessary.

This implementation is more comprehensive than what is implemented in
package validator, where package validator will have false negatives
that this implementation will properly flag.