(#38) Add rule for validating ID length
This implements the note rule CPMR0069, that verifies that the length of
the package identifier is not longer than 20 characters without a dash
between the words.

This implementation is more restrict than what is currently implemented
on Package Validator as it will also consider packages that have certain
extension, just by removing this extension first and it will also verify
that any section between dashes is not longer than 20 characters.