(maint) Update GRM configuration
We are consolidating labels that are in use across our repositories, so this commit brings the GRM configuration up to date.  Namely, we are using Enhancement now, where we might have used Improvement or Feature before.  Also, removed C4B Feature as this has never been used. Finally, update the Internal Refactoring label to use the new `Area - Refactor` label.

Signed-off-by: Gary Ewan Park <gep13@gep13.co.uk>